Move TUI renderer to own file
